+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_.cs
@@ -1,9 +1,9 @@
 namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
 
-/// <summary>   A session base methods. </summary>
+/// <summary>   A session base extension methods. </summary>
 /// <remarks>   2025-04-10. </remarks>
 [CLSCompliant( false )]
-public static partial class Methods
+public static partial class SessionBaseExtensionMethods
 {
     /// <summary>   Trace last action. </summary>
     /// <remarks>   2025-04-10. </remarks>