+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_export.cs
@@ -3,7 +3,7 @@ using cc.isr.VI.Pith;
 
 namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
 
-public static partial class Methods
+public static partial class SessionBaseExtensionMethods
 {
     /// <summary>   A <see cref="Pith.SessionBase"/> extension method that fetches and exports a script to file. </summary>
     /// <remarks>   2025-04-10. </remarks>
@@ -51,18 +51,14 @@ public static partial class Methods
     ///                                             null. </exception>
     /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
     /// <param name="reader">       The reader. </param>
+    /// <param name="writer">       The writer. </param>
     /// <param name="fromEol">      (Optional) end-of-line to change. </param>
     /// <param name="toEol">        (Optional) end-of-line to change to. </param>
-    /// <param name="overWrite">    (Optional) [false] True to over write. </param>
     /// <returns>   The EOL converted. </returns>
-    public static string ConvertEol( this TextReader reader, string fromEol = "\n", string toEol = "\r\n", bool overWrite = false )
+    public static void ConvertEol( this TextReader reader, TextWriter writer, string fromEol = "\n", string toEol = "\r\n" )
     {
         if ( reader == null )
             throw new ArgumentNullException( nameof( reader ) );
-
-        StringBuilder sb = new();
-        using TextWriter? writer = new System.IO.StringWriter( sb )
-            ?? throw new System.IO.FileNotFoundException( $"Failed creating a writer for the string builder" );
 
         string? line = "";
         bool? isExpectedLineFormat = null;
@@ -85,7 +81,6 @@ public static partial class Methods
                 writer.Write( line );
             }
         }
-        return sb.ToString();
     }
 
     /// <summary>   A <see cref="TextReader"/> extension method that exports script to file preserving the Windows end-of-line format. </summary>
@@ -182,7 +177,7 @@ public static partial class Methods
                 $"The exported script source file '{filePath}' length {fileInfo.Length} is smaller that the script source length {scriptSource.Length}." );
     }
 
-    /// <summary>   A string extension method that compress script. </summary>
+    /// <summary>   A <see cref="string"/> extension method that compress script source to a file. </summary>
     /// <remarks>   2025-04-14. </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
     ///                                                 are null. </exception>
@@ -191,19 +186,19 @@ public static partial class Methods
     /// <exception cref="FileNotFoundException">        Thrown when the requested file is not
     ///                                                 present. </exception>
     /// <param name="scriptSource"> The script source. </param>
-    /// <param name="filePath">     Full pathname of the file. </param>
+    /// <param name="toFilePath">   the destination file path. </param>
     /// <param name="fromEol">      (Optional) end-of-line to change. </param>
     /// <param name="toEol">        (Optional) end-of-line to change to. </param>
     /// <param name="overWrite">    (Optional) [false] True to over write. </param>
-    public static void CompressScript( this string scriptSource, string filePath, string fromEol = "\n", string toEol = "\r\n", bool overWrite = false )
+    public static void CompressScript( this string scriptSource, string toFilePath, string fromEol = "\n", string toEol = "\r\n", bool overWrite = false )
     {
         if ( string.IsNullOrWhiteSpace( scriptSource ) )
             throw new ArgumentNullException( nameof( scriptSource ) );
-        if ( string.IsNullOrWhiteSpace( filePath ) )
-            throw new ArgumentNullException( nameof( filePath ) );
+        if ( string.IsNullOrWhiteSpace( toFilePath ) )
+            throw new ArgumentNullException( nameof( toFilePath ) );
 
-        if ( !overWrite && System.IO.File.Exists( filePath ) )
-            throw new InvalidOperationException( $"The script source cannot be exported because the file '{filePath}' exists." );
+        if ( !overWrite && System.IO.File.Exists( toFilePath ) )
+            throw new InvalidOperationException( $"The script source cannot be exported because the file '{toFilePath}' exists." );
 
         if ( !(string.IsNullOrEmpty( fromEol ) || string.IsNullOrEmpty( toEol )) )
         {
@@ -211,11 +206,69 @@ public static partial class Methods
             using TextReader? reader = new System.IO.StringReader( scriptSource )
                 ?? throw new System.IO.FileNotFoundException( $"Failed creating a reader from the script source" );
 
+            StringBuilder sb = new();
+            using TextWriter? writer = new System.IO.StringWriter( sb )
+                ?? throw new System.IO.FileNotFoundException( $"Failed creating a writer for the string builder" );
+
             // convert line ending to windows format.
-            scriptSource = reader.ConvertEol( fromEol, toEol, overWrite );
+            reader.ConvertEol( writer, fromEol, toEol );
+
+            scriptSource = sb.ToString();
         }
 
         // compress and export the source to the file as is.
-        System.IO.File.WriteAllText( filePath, ScriptCompressor.Compress( scriptSource ) );
+        System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( scriptSource ) );
     }
+
+    /// <summary>   A <see cref="string"/> extension method that compress a script file to a destination file. </summary>
+    /// <remarks>   2025-04-15. </remarks>
+    /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
+    ///                                                 are null. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <exception cref="FileNotFoundException">        Thrown when the requested file is not
+    ///                                                 present. </exception>
+    /// <param name="fromFilePath"> The source file path. </param>
+    /// <param name="toFilePath">   the destination file path. </param>
+    /// <param name="fromEol">      (Optional) end-of-line to change. </param>
+    /// <param name="toEol">        (Optional) end-of-line to change to. </param>
+    /// <param name="overWrite">    (Optional) [false] True to over write. </param>
+    public static void CompressScriptFile( this string fromFilePath, string toFilePath, string fromEol = "\n", string toEol = "\r\n", bool overWrite = false )
+    {
+        if ( string.IsNullOrWhiteSpace( fromFilePath ) )
+            throw new ArgumentNullException( nameof( fromFilePath ) );
+        if ( string.IsNullOrWhiteSpace( toFilePath ) )
+            throw new ArgumentNullException( nameof( toFilePath ) );
+
+        if ( !overWrite && System.IO.File.Exists( toFilePath ) )
+            throw new InvalidOperationException( $"The script source cannot be exported because the file '{toFilePath}' exists." );
+
+        string scriptSource;
+
+        if ( !(string.IsNullOrEmpty( fromEol ) || string.IsNullOrEmpty( toEol )) )
+        {
+            // if required, convert the line endings to windows format.
+            using TextReader? reader = new System.IO.StreamReader( fromFilePath )
+                ?? throw new System.IO.FileNotFoundException( $"Failed creating a reader from the '{fromFilePath}' file." );
+
+            StringBuilder sb = new();
+            using TextWriter? writer = new System.IO.StringWriter( sb )
+                ?? throw new System.IO.FileNotFoundException( $"Failed creating a writer for the string builder" );
+
+            // convert line ending to windows format.
+            reader.ConvertEol( writer, fromEol, toEol );
+
+            // convert line ending to windows format.
+            scriptSource = sb.ToString();
+        }
+        else
+        {
+            // read all text as is.
+            scriptSource = System.IO.File.ReadAllText( fromFilePath );
+        }
+
+        // compress and export the source to the file as is.
+        System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( scriptSource ) );
+    }
+
 }