+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_run.cs
@@ -2,7 +2,7 @@ using cc.isr.VI.Pith;
 
 namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
 
-public static partial class Methods
+public static partial class SessionBaseExtensionMethods
 {
     /// <summary>
     /// A <see cref="Pith.SessionBase"/> extension method that query if the specified script is